+++ src/main/java/org/zerock/wego/controller/MainController.java
@@ -7,13 +7,13 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.zerock.wego.domain.FavoriteVO;
-import org.zerock.wego.domain.MountainInfoViewVO;
-import org.zerock.wego.domain.RecruitmentViewVO;
+import org.zerock.wego.domain.SanInfoViewVO;
+import org.zerock.wego.domain.PartyViewVO;
 import org.zerock.wego.domain.ReviewViewVO;
 import org.zerock.wego.exception.ControllerException;
 import org.zerock.wego.service.FavoriteService;
-import org.zerock.wego.service.MountainInfoService;
-import org.zerock.wego.service.RecruitmentService;
+import org.zerock.wego.service.SanInfoService;
+import org.zerock.wego.service.PartyService;
 import org.zerock.wego.service.ReviewService;
 import org.zerock.wego.service.SearchService;
 
@@ -26,8 +26,8 @@ import lombok.extern.log4j.Log4j2;
 @RequestMapping("/") // BASE URL
 @Controller
 public class MainController {
-	private MountainInfoService mountainInfoService;
-	private RecruitmentService recruitmentService;
+	private SanInfoService mountainInfoService;
+	private PartyService recruitmentService;
 	private ReviewService reviewService;
 	private FavoriteService favoriteService;
 
@@ -41,11 +41,11 @@ public class MainController {
 			model.addAttribute("favoriteList", favoriteList);
 
 			// =========== 산정보 ===========
-			Set<MountainInfoViewVO> mountainInfoList = this.mountainInfoService.getRandom10List();
+			Set<SanInfoViewVO> mountainInfoList = this.mountainInfoService.getRandom10List();
 			model.addAttribute("mountainInfoList", mountainInfoList);
 
 			// =========== 모집글 ===========
-			Set<RecruitmentViewVO> recruitmentList = this.recruitmentService.getRandom10List();
+			Set<PartyViewVO> recruitmentList = this.recruitmentService.getRandom10List();
 			model.addAttribute("recruitmentList", recruitmentList);
 
 			// =========== 후기글 ===========
