+++ src/main/java/org/zerock/wego/controller/SearchController.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.zerock.wego.domain.MountainInfoViewVO;
-import org.zerock.wego.domain.RecruitmentViewVO;
+import org.zerock.wego.domain.SanInfoViewVO;
+import org.zerock.wego.domain.PartyViewVO;
 import org.zerock.wego.domain.ReviewViewVO;
 import org.zerock.wego.exception.ControllerException;
 import org.zerock.wego.service.SearchService;
@@ -30,11 +30,11 @@ public class SearchController {
 
 		try {
 			// =========== 산정보 ===========
-			Set<MountainInfoViewVO> mountainInfoList = this.service.selectSearchMountainInfo(search);
+			Set<SanInfoViewVO> mountainInfoList = this.service.selectSearchMountainInfo(search);
 			model.addAttribute("mountainInfoList", mountainInfoList);
 
 			// =========== 모집글 ===========
-			Set<RecruitmentViewVO> recruitmentList = this.service.selectSearchRecruitment(search);
+			Set<PartyViewVO> recruitmentList = this.service.selectSearchRecruitment(search);
 			model.addAttribute("recruitmentList", recruitmentList);
 
 			// =========== 후기글 ===========
