--- src/main/java/org/zerock/wego/service/No.java
@@ -1,5 +1,0 @@
-package org.zerock.wego.service;
-
-public class No {
-
-}
