+++ src/main/java/org/zerock/wego/service/PartyService.java
@@ -1,46 +1,28 @@
 package org.zerock.wego.service;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.zerock.wego.domain.RecruitmentDTO;
-import org.zerock.wego.domain.RecruitmentViewVO;
+import org.zerock.wego.domain.PartyDTO;
+import org.zerock.wego.domain.PartyViewVO;
 import org.zerock.wego.exception.ServiceException;
-import org.zerock.wego.mapper.RecruitmentMapper;
+import org.zerock.wego.mapper.PartyMapper;
 
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
-@ToString
+
 @Log4j2
-@NoArgsConstructor
+@RequiredArgsConstructor
 
 @Service
-public class RecruitmentServiceImpl implements RecruitmentService, InitializingBean { // POJO(상속X)
+public class PartyService {
+	
+	private final PartyMapper mapper;
 
-	@Setter(onMethod_ = { @Autowired })
-	private RecruitmentMapper mapper;
 
-	@Override
-	public void afterPropertiesSet() throws ServiceException { // 1회성 전처리
-		log.trace("afterPropertiesSet() invoked.");
-
-		try {
-			Objects.requireNonNull(this.mapper);
-			log.info("this.mapper: {}", this.mapper);
-		} catch (Exception e) {
-			throw new ServiceException(e);
-		} // try-catch
-	} // afterPropertiesSet
-
-	@Override
-	public List<RecruitmentViewVO> getList() throws ServiceException {
+	public List<PartyViewVO> getList() throws ServiceException {
 		log.trace("getList() invoked.");
 
 		try {
@@ -50,8 +32,7 @@ public class RecruitmentServiceImpl implements RecruitmentService, InitializingB
 		} // try-catch
 	} // getList
 
-	@Override
-	public Set<RecruitmentViewVO> getRandom10List() throws ServiceException {
+	public Set<PartyViewVO> getRandom10List() throws ServiceException {
 		log.trace("getRandom10List() invoked.");
 
 		try {
@@ -61,8 +42,7 @@ public class RecruitmentServiceImpl implements RecruitmentService, InitializingB
 		} // try-catch
 	} // getList
 
-	@Override
-	public RecruitmentViewVO get(Integer sanPartyId) throws ServiceException {
+	public PartyViewVO get(Integer sanPartyId) throws ServiceException {
 		log.trace("get({}) invoked.", sanPartyId);
 
 		try {
@@ -72,7 +52,6 @@ public class RecruitmentServiceImpl implements RecruitmentService, InitializingB
 		} // try-catch
 	} // get
 
-	@Override
 	public boolean remove(Integer sanPartyId) throws ServiceException {
 		log.trace("remove({}) invoked.", sanPartyId);
 
@@ -83,8 +62,7 @@ public class RecruitmentServiceImpl implements RecruitmentService, InitializingB
 		} // try-catch
 	} // remove
 
-	@Override
-	public boolean register(RecruitmentDTO dto) throws ServiceException {
+	public boolean register(PartyDTO dto) throws ServiceException {
 		log.trace("register({}) invoked.", dto);
 
 		try {
@@ -94,8 +72,7 @@ public class RecruitmentServiceImpl implements RecruitmentService, InitializingB
 		} // try-catch
 	} // register
 
-	@Override
-	public boolean modify(RecruitmentDTO dto) throws ServiceException {
+	public boolean modify(PartyDTO dto) throws ServiceException {
 		log.trace("modify({}) invoked.", dto);
 
 		try {
