+++ src/main/java/org/zerock/wego/service/ReviewService.java
@@ -1,31 +1,91 @@
 package org.zerock.wego.service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 
+import org.springframework.stereotype.Service;
 import org.zerock.wego.domain.ReviewDTO;
 import org.zerock.wego.domain.ReviewViewVO;
 import org.zerock.wego.exception.ServiceException;
+import org.zerock.wego.mapper.ReviewMapper;
 
-public interface ReviewService {
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
-//	// 게시물 전체목록 조회(LIST)
-	public abstract List<ReviewViewVO> getList() throws ServiceException;
 
-	// 10개의 게시물 랜덤 조회(LIST) - 메인페이지용
-	public abstract Set<ReviewViewVO> getRandom10List() throws ServiceException;
+@Log4j2
+@RequiredArgsConstructor
 
-	// 특정 후기글 상세조회(READ)
-	public abstract ReviewViewVO getReviewByReviewId(Integer sanReviewId) throws ServiceException;
+@Service
+public class ReviewService {
 
-	// 특정 후기글 삭제(DELETE)
-	public abstract boolean isReviewRemove(Integer sanReviewId) throws ServiceException;
+	private final ReviewMapper mapper;
 
-	// 새로운 게시물 등록(CREATE)
-	public abstract boolean register(ReviewDTO dto) throws ServiceException;
+	public List<ReviewViewVO> getList() throws ServiceException {
+		log.trace("getList() invoked.");
+		
+		try {
+			return this.mapper.selectAll();
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // getList
 
-	// 기존 게시물 수정(UPDATE)
-	public abstract boolean modify(ReviewDTO dto) throws ServiceException;
+	public Set<ReviewViewVO> getRandom10List() throws ServiceException {
+		log.trace("getRandom10List() invoked.");
+		
+		try {
+			return this.mapper.selectRandom10();
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // getList
 
-} // end interface
-	
+	public ReviewViewVO getReviewByReviewId(Integer sanReviewId) throws ServiceException {
+		log.trace("get({}) invoked.", sanReviewId);	
+		
+		try {
+			ReviewViewVO review = this.mapper.selectReviewByReviewId(sanReviewId);
+			Objects.requireNonNull(review);
+
+			
+			return review;
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // get
+
+	public boolean isReviewRemove(Integer sanReviewId) throws ServiceException {
+		log.trace("remove({}) invoked.", sanReviewId);
+		
+		try {
+			
+			return this.mapper.deleteReviewByReviewId(sanReviewId) == 1;
+			
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // remove
+
+	public boolean register(ReviewDTO dto) throws ServiceException {
+		log.trace("register({}) invoked.");
+		
+		try {
+			return this.mapper.insert(dto) == 1;
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // register
+
+	public boolean modify(ReviewDTO dto) throws ServiceException {
+		log.trace("modify({}) invoked.");
+		
+		try {
+			return this.mapper.update(dto) == 1;
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // modify
+
+}// end class
