+++ src/main/java/org/zerock/wego/service/SanInfoService.java
@@ -1,0 +1,64 @@
+package org.zerock.wego.service;
+
+import java.util.List;
+import java.util.Set;
+
+import org.springframework.stereotype.Service;
+import org.zerock.wego.domain.SanInfoViewVO;
+import org.zerock.wego.exception.ServiceException;
+import org.zerock.wego.mapper.SanInfoMapper;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+
+
+@Log4j2
+@RequiredArgsConstructor
+
+@Service
+public class SanInfoService { // POJO(상속X)
+
+	private final SanInfoMapper mapper;
+
+
+	public List<SanInfoViewVO> getList() throws ServiceException {
+		log.trace("getList() invoked.");
+		
+		try {
+			return this.mapper.selectAll();
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // getList
+	
+	public Set<SanInfoViewVO> getRandom10List() throws ServiceException {
+		log.trace("getRandom10List() invoked.");
+		
+		try {
+			return this.mapper.selectRandom10();
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // getList
+	
+	public SanInfoViewVO get(Integer sanInfoId) throws ServiceException {
+		log.trace("get({}) invoked.", sanInfoId);	
+		
+		try {
+			return this.mapper.select(sanInfoId);
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // get
+	
+	public Integer selectSanName(String sanName) throws ServiceException {
+		log.trace("selectSanName({}) invoked.", sanName);
+		
+		try {
+			return this.mapper.selectSanName(sanName);
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // selectSanName
+	
+} // end class
