+++ src/main/java/org/zerock/wego/service/SearchService.java
@@ -2,16 +2,54 @@ package org.zerock.wego.service;
 
 import java.util.Set;
 
-import org.zerock.wego.domain.MountainInfoViewVO;
-import org.zerock.wego.domain.RecruitmentViewVO;
+import org.springframework.stereotype.Service;
+import org.zerock.wego.domain.PartyViewVO;
 import org.zerock.wego.domain.ReviewViewVO;
+import org.zerock.wego.domain.SanInfoViewVO;
 import org.zerock.wego.exception.ServiceException;
+import org.zerock.wego.mapper.SearchMapper;
 
-public interface SearchService {
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
-	// 통합검색결과 조회(산 => 산이름, 상세 / 모집, 후기 => 제목, 내용)
-	public abstract Set<MountainInfoViewVO> selectSearchMountainInfo(String search) throws ServiceException;
-	public abstract Set<RecruitmentViewVO> selectSearchRecruitment(String search) throws ServiceException;
-	public abstract Set<ReviewViewVO> selectSearchReview(String search) throws ServiceException;
+
+@Log4j2
+@RequiredArgsConstructor
+
+@Service
+public class SearchService {
+
+	private final SearchMapper mapper;
+
+	
+	public Set<SanInfoViewVO> selectSearchMountainInfo(String search) throws ServiceException {
+		log.trace("selectSearchMountainInfo({}) invoked.", search);
+		
+		try {
+			return this.mapper.selectSearchMountainInfo(search);
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // selectSearchMountainInfo
+
+	public Set<PartyViewVO> selectSearchRecruitment(String search) throws ServiceException {
+		log.trace("selectSearchRecruitment({}) invoked.", search);
+		
+		try {
+			return this.mapper.selectSearchRecruitment(search);
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // selectSearchRecruitment
+
+	public Set<ReviewViewVO> selectSearchReview(String search) throws ServiceException {
+		log.trace("selectSearchReview({}) invoked.", search);
+		
+		try {
+			return this.mapper.selectSearchReview(search);
+		} catch (Exception e) {
+			throw new ServiceException(e);
+		} // try-catch
+	} // selectSearchReview
 
 } // end interface
