+++ src/main/resources/org/zerock/wego/mapper/PartyMapper.xml
